requirements.txt
pyspark==3.5.1
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.1.1
datetime==5.5
psycopg2==2.9.9
dbt-core==1.8.1
dbt-postgres==1.8.1